refactor(remediationorchestrator): extract pre_remediation_hash.go from reconciler.go
Phase 3 file split: moves resolveDualTargets, formatRemediationTargetString,
CapturePreRemediationHash, and resolveConfigMapHashes (DD-EM-002/EM-003 hash
capture logic) into their own file. Pure structural move, no behavior change:
376 Ginkgo specs still pass, same package, no import-path changes.
